Fix the fails() by repurposing it as an assertion
